requirements.txt
django
psycopg2
tzdata
djangorestframework
markdown
django-filter